Refactor code to introduce TrainingMapper
In this commit, refactored code encapsulates business logic in TrainingMapper, moving from TrainingService and TrainingController. This was achieved in order to make the main Controller and Service components cleaner and easier to read. Import statements in several classes were rearranged, and minor changes were made to method structure to further improve clarity.